requirements.txt
# Fetch.ai uAgents Framework (Required for hackathon)
uagents>=0.12.0

# AI/LLM Integration (Required for hackathon)
anthropic>=0.18.0

# Legacy OpenAI support (optional for demo mode)
openai>=1.0.0

# Additional utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
httpx>=0.25.0

# Web Backend
fastapi>=0.104.0
uvicorn>=0.24.0
flask>=3.0.0
flask-socketio>=5.3.0
python-socketio>=5.10.0

# Web UI
streamlit>=1.28.0
